@@ -1,18 +1,17 @@
 name: flutter_widget_from_html
-version: 0.6.0-rc.2021030202
+version: 0.6.0-rc.2021030401
 description: Flutter package for widget tree building from html that supports hyperlink, image, nested list, etc.
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  # flutter_widget_from_html_core
-  flutter: ">=1.20.0"
-  sdk: ">=2.7.0 <3.0.0"
+  flutter: ">=2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  cached_network_image: ^2.2.0+1
   flutter:
     sdk: flutter
   flutter_widget_from_html_core: ^0.6.0-0
+  fwfh_cached_network_image: ^0.6.0-0
   fwfh_chewie: ^0.6.0-0
   fwfh_svg: ^0.6.0-0
   fwfh_url_launcher: ^0.6.0-0
@@ -22,6 +21,8 @@ dependencies:
 dependency_overrides:
   flutter_widget_from_html_core:
     path: ../core
+  fwfh_cached_network_image:
+    path: ../fwfh_cached_network_image
   fwfh_chewie:
     path: ../fwfh_chewie
   fwfh_svg:
@@ -30,7 +31,6 @@ dependency_overrides:
     path: ../fwfh_url_launcher
   fwfh_webview:
     path: ../fwfh_webview
-  wakelock: ^0.4.0
 
 dev_dependencies:
   flutter_test:
