@@ -18,8 +18,7 @@ class Golden extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final baseUrl = Uri.parse('https://www.w3schools.com/html/');
-    final withEnhanced =
-        RegExp(r'(^(A|HR|IFRAME|SVG|VIDEO)$|colspan|rowspan)').hasMatch(name);
+    final withEnhanced = RegExp(r'^(IFRAME|SVG|VIDEO)$').hasMatch(name);
 
     final children = <Widget>[
       Text(html),
