@@ -6,16 +6,12 @@ const kTagA = 'a';
 
 class TagA {
   final WidgetFactory wf;
-  final Color? Function() colorCallback;
 
-  TagA(this.wf, this.colorCallback);
+  TagA(this.wf);
 
   BuildOp get buildOp => BuildOp(
         defaultStyles: (_) {
           final styles = {kCssTextDecoration: kCssTextDecorationUnderline};
-
-          final color = colorCallback.call();
-          if (color != null) styles[kCssColor] = convertColorToHex(color);
 
           return styles;
         },
