@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
@@ -10,7 +11,7 @@ Future<String> explain(WidgetTester t, HtmlWidget hw) =>
 
 void main() {
   group('buildAsync', () {
-    final explain = (WidgetTester tester, String html, bool buildAsync) =>
+    final explain = (WidgetTester tester, String html, bool? buildAsync) =>
         tester.runAsync(() => helper.explain(tester, null,
             hw: HtmlWidget(
               html,
@@ -41,8 +42,8 @@ void main() {
     final explain = (
       WidgetTester tester,
       String html, {
-      AsyncWidgetBuilder<Widget> buildAsyncBuilder,
-      bool withData,
+      AsyncWidgetBuilder<Widget>? buildAsyncBuilder,
+      required bool withData,
     }) =>
         tester.runAsync(() => helper.explain(tester, null,
             buildFutureBuilderWithData: withData,
@@ -60,7 +61,8 @@ void main() {
         expect(explained, equals('[FutureBuilder:[RichText:(:$html)]]'));
       });
 
-      testWidgets('renders indicator', (WidgetTester tester) async {
+      testWidgets('renders CircularProgressIndicator', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.android;
         final html = 'Foo';
         final explained = await explain(tester, html, withData: false);
         expect(
@@ -70,13 +72,28 @@ void main() {
                 '[Padding:(8,8,8,8),child='
                 '[CircularProgressIndicator]'
                 ']]]'));
+        debugDefaultTargetPlatformOverride = null;
+      });
+
+      testWidgets('renders CupertinoActivityIndicator', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+        final html = 'Foo';
+        final explained = await explain(tester, html, withData: false);
+        expect(
+            explained,
+            equals('[FutureBuilder:'
+                '[Center:child='
+                '[Padding:(8,8,8,8),child='
+                '[CupertinoActivityIndicator]'
+                ']]]'));
+        debugDefaultTargetPlatformOverride = null;
       });
     });
 
     group('custom', () {
       final buildAsyncBuilder =
           (BuildContext _, AsyncSnapshot<Widget> snapshot) =>
-              snapshot.hasData ? snapshot.data : Text('No data');
+              snapshot.data ?? Text('No data');
 
       testWidgets('renders data', (WidgetTester tester) async {
         final html = 'Foo';
@@ -107,11 +124,11 @@ void main() {
       WidgetTester tester,
       String html,
       bool enableCaching, {
-      Uri baseUrl,
-      bool buildAsync,
+      Uri? baseUrl,
+      bool? buildAsync,
       Color hyperlinkColor = const Color.fromRGBO(0, 0, 255, 1),
-      RebuildTriggers rebuildTriggers,
-      TextStyle textStyle,
+      RebuildTriggers? rebuildTriggers,
+      TextStyle? textStyle,
       bool webView = false,
       bool webViewJs = true,
     }) =>
@@ -124,12 +141,12 @@ void main() {
               hyperlinkColor: hyperlinkColor,
               key: helper.hwKey,
               rebuildTriggers: rebuildTriggers,
-              textStyle: textStyle,
+              textStyle: textStyle ?? const TextStyle(),
               webView: webView,
               webViewJs: webViewJs,
             ));
 
-    final _expect = (Widget built1, Widget built2, Matcher matcher) {
+    final _expect = (Widget? built1, Widget? built2, Matcher matcher) {
       final widget1 = (built1 as TshWidget).child;
       final widget2 = (built2 as TshWidget).child;
       expect(widget1 == widget2, matcher);
@@ -442,17 +459,6 @@ void main() {
       expect(explained, equals('[GestureDetector:child=[Text:$webViewSrc]]'));
     });
 
-    testWidgets('renders null value', (WidgetTester tester) async {
-      final explained = await explain(
-          tester,
-          HtmlWidget(
-            html,
-            key: helper.hwKey,
-            webView: null,
-          ));
-      expect(explained, equals('[GestureDetector:child=[Text:$webViewSrc]]'));
-    });
-
     group('webViewDebuggingEnabled', () {
       testWidgets('renders true value', (WidgetTester tester) async {
         final explained = await explain(
@@ -481,24 +487,6 @@ void main() {
               key: helper.hwKey,
               webView: true,
               webViewDebuggingEnabled: false,
-            ));
-        expect(
-            explained,
-            equals('[WebView:'
-                'url=$webViewSrc,'
-                'aspectRatio=$webViewDefaultAspectRatio,'
-                'autoResize=true'
-                ']'));
-      });
-
-      testWidgets('renders null value', (WidgetTester tester) async {
-        final explained = await explain(
-            tester,
-            HtmlWidget(
-              html,
-              key: helper.hwKey,
-              webView: true,
-              webViewDebuggingEnabled: null,
             ));
         expect(
             explained,
@@ -546,24 +534,6 @@ void main() {
                 'js=false'
                 ']'));
       });
-
-      testWidgets('renders null value', (WidgetTester tester) async {
-        final explained = await explain(
-            tester,
-            HtmlWidget(
-              html,
-              key: helper.hwKey,
-              webView: true,
-              webViewJs: null,
-            ));
-        expect(
-            explained,
-            equals('[WebView:'
-                'url=$webViewSrc,'
-                'aspectRatio=$webViewDefaultAspectRatio,'
-                'js=false'
-                ']'));
-      });
     });
 
     group('webViewMediaPlaybackAlwaysAllow', () {
@@ -594,24 +564,6 @@ void main() {
               key: helper.hwKey,
               webView: true,
               webViewMediaPlaybackAlwaysAllow: false,
-            ));
-        expect(
-            explained,
-            equals('[WebView:'
-                'url=$webViewSrc,'
-                'aspectRatio=$webViewDefaultAspectRatio,'
-                'autoResize=true'
-                ']'));
-      });
-
-      testWidgets('renders null value', (WidgetTester tester) async {
-        final explained = await explain(
-            tester,
-            HtmlWidget(
-              html,
-              key: helper.hwKey,
-              webView: true,
-              webViewMediaPlaybackAlwaysAllow: null,
             ));
         expect(
             explained,
