@@ -22,31 +22,30 @@ void main() {
 
     final videoPlayer = 'TshWidget\n'
         '└WidgetPlaceholder<Widget>(VideoPlayer)\n'
-        ' └VideoPlayer()\n'
-        '  └_VideoPlayerWidget(state: _VideoPlayerState)\n'
-        '   └AspectRatio(aspectRatio: 1.8)\n'
-        '    └Chewie(state: ChewieState)\n'
-        '     └_ChewieControllerProvider\n'
-        '      └PlayerWithControls()\n'
-        '       └Center(alignment: center)\n'
-        '        └SizedBox(width: 800.0, height: 600.0)\n'
-        '         └AspectRatio(aspectRatio: 1.3)\n'
-        '          └Stack(alignment: topStart, fit: loose)\n'
-        '           ├Container\n'
-        '           │└LimitedBox(maxWidth: 0.0, maxHeight: 0.0)\n'
-        '           │ └ConstrainedBox(BoxConstraints(biggest))\n'
-        '           ├Center(alignment: center)\n'
-        '           │└AspectRatio(aspectRatio: 1.8)\n'
-        '           │ └VideoPlayer(state: _VideoPlayerState)\n'
-        '           │  └Container\n'
-        '           │   └LimitedBox(maxWidth: 0.0, maxHeight: 0.0)\n'
-        '           │    └ConstrainedBox(BoxConstraints(biggest))\n'
-        '           ├Container\n'
-        '           │└LimitedBox(maxWidth: 0.0, maxHeight: 0.0)\n'
-        '           │ └ConstrainedBox(BoxConstraints(biggest))\n'
-        '           └Container\n'
-        '            └LimitedBox(maxWidth: 0.0, maxHeight: 0.0)\n'
-        '             └ConstrainedBox(BoxConstraints(biggest))\n'
+        ' └VideoPlayer(state: _VideoPlayerState)\n'
+        '  └AspectRatio(aspectRatio: 1.8)\n'
+        '   └Chewie(state: ChewieState)\n'
+        '    └_ChewieControllerProvider\n'
+        '     └PlayerWithControls()\n'
+        '      └Center(alignment: center)\n'
+        '       └SizedBox(width: 800.0, height: 600.0)\n'
+        '        └AspectRatio(aspectRatio: 1.3)\n'
+        '         └Stack(alignment: topStart, fit: loose)\n'
+        '          ├Container\n'
+        '          │└LimitedBox(maxWidth: 0.0, maxHeight: 0.0)\n'
+        '          │ └ConstrainedBox(BoxConstraints(biggest))\n'
+        '          ├Center(alignment: center)\n'
+        '          │└AspectRatio(aspectRatio: 1.8)\n'
+        '          │ └VideoPlayer(state: _VideoPlayerState)\n'
+        '          │  └Container\n'
+        '          │   └LimitedBox(maxWidth: 0.0, maxHeight: 0.0)\n'
+        '          │    └ConstrainedBox(BoxConstraints(biggest))\n'
+        '          ├Container\n'
+        '          │└LimitedBox(maxWidth: 0.0, maxHeight: 0.0)\n'
+        '          │ └ConstrainedBox(BoxConstraints(biggest))\n'
+        '          └Container\n'
+        '           └LimitedBox(maxWidth: 0.0, maxHeight: 0.0)\n'
+        '            └ConstrainedBox(BoxConstraints(biggest))\n'
         '\n';
 
     testWidgets('renders video player (Android)', (tester) async {
@@ -70,13 +69,12 @@ void main() {
           explained,
           equals('TshWidget\n'
               '└WidgetPlaceholder<Widget>(VideoPlayer)\n'
-              ' └VideoPlayer()\n'
-              '  └_VideoPlayerWidget(state: _PlaceholderState)\n'
-              '   └AspectRatio(aspectRatio: 1.8)\n'
-              '    └DecoratedBox(bg: BoxDecoration(color: Color(0x7f000000)))\n'
-              '     └Center(alignment: center)\n'
-              '      └Text("platform=TargetPlatform.linux")\n'
-              '       └RichText(text: "platform=TargetPlatform.linux")\n'
+              ' └VideoPlayer(state: _PlaceholderState)\n'
+              '  └AspectRatio(aspectRatio: 1.8)\n'
+              '   └DecoratedBox(bg: BoxDecoration(color: Color(0x7f000000)))\n'
+              '    └Center(alignment: center)\n'
+              '     └Text("platform=TargetPlatform.linux")\n'
+              '      └RichText(text: "platform=TargetPlatform.linux")\n'
               '\n'));
       debugDefaultTargetPlatformOverride = null;
     });
